+++ controller/contentController.js
@@ -1,5 +1,6 @@
 const { body, validationResult } = require("express-validator");
 const Content = require("../model/Content");
+const Product = require("../model/Product");
 
 exports.validate = [
   body("name")
@@ -78,6 +79,14 @@ exports.update = async (req, res) => {
 exports.deleteById = async (req, res) => {
   try {
     const { id } = req.params;
+    // Prevent deletion if referenced by any product
+    const productUsing = await Product.findOne({ content: id });
+    if (productUsing) {
+      return res.status(400).json({
+        success: false,
+        message: "Cannot delete: Content is in use by one or more products.",
+      });
+    }
     const deleted = await Content.findByIdAndDelete(id);
     if (!deleted) {
       return res.status(404).json({ success: false, message: "Not found" });
