+++ controller/subfinishController.js
@@ -1,6 +1,7 @@
 const { body, validationResult } = require("express-validator");
 const Subfinish = require("../model/Subfinish");
 const Finish = require("../model/Finish");
+const Product = require("../model/Product");
 
 exports.validate = [
   body("name")
@@ -92,6 +93,14 @@ exports.update = async (req, res) => {
 exports.deleteById = async (req, res) => {
   try {
     const { id } = req.params;
+    // Prevent deletion if referenced by any product
+    const productUsing = await Product.findOne({ subfinish: id });
+    if (productUsing) {
+      return res.status(400).json({
+        success: false,
+        message: "Cannot delete: Subfinish is in use by one or more products.",
+      });
+    }
     const deleted = await Subfinish.findByIdAndDelete(id);
     if (!deleted) {
       return res.status(404).json({ success: false, message: "Not found" });
